backend/requirements.txt
# Core AI/ML dependencies
google-genai>=0.3.0

# Voice/Speech APIs
elevenlabs>=1.0.0

# Web framework and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.5.0

# Environment management
python-dotenv>=1.0.0

# Audio processing
pydub>=0.25.1
